pyproject.toml
[tool.poetry]
name = "poetry"
version = "0.1.0"
description = "Poetry django"
authors = ["samikshakhadka <swabhimankhadka2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
psycopg = "^3.2.1"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.3"
celery = "^5.4.0"
drf-yasg = "^1.21.7"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
django-stubs = "^5.0.4"
mypy = "^1.11.1"
isort = "^5.13.2"
black = "^24.8.0"
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"